# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TorqueLane API",
        "version": "0.1.0",
        "description": "TorqueLane API. Errors always use the envelope { \"error\": { \"code\", \"message\", \"details\"? } }; clients branch on `code`. Records outside the tenant scope of the caller answer 404 not_found, exactly like records that do not exist."
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Relative to the environment host, e.g. https://api-staging.<domain>"
        }
    ],
    "paths": {
...
        "/customer-accounts/{customerAccount}/consents/current": {
            "get": {
                "operationId": "customer-accounts.consents.current",
                "description": "The latest decision per purpose, for the account and for each contact.",
                "summary": "Current consent",
                "tags": [
                    "Consent"
                ],
                "parameters": [
                    {
                        "name": "customerAccount",
                        "in": "path",
                        "required": true,
                        "description": "The customer account ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ConsentStateResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ConsentStateResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Error.forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/Error.not_found"
                    },
                    "401": {
                        "$ref": "#/components/responses/Error.unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ConsentStateResource": {
                "type": "object",
                "properties": {
                    "account": {
                        "type": "array",
                        "items": {}
                    },
                    "contacts": {
                        "type": "string"
                    }
                },
                "required": [
                    "account",
                    "contacts"
                ],
                "title": "ConsentStateResource"
            },
...
        },
        "responses": {
            "Error.forbidden": {
                "description": "This action is unauthorized.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "enum": [
                                                "forbidden"
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable; never branch on it."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional machine-readable context."
                                        }
                                    },
                                    "required": [
                                        "code",
                                        "message"
                                    ]
                                }
                            },
                            "required": [
                                "error"
                            ]
                        }
                    }
                }
            },
...
            "Error.unauthenticated": {
                "description": "Unauthenticated.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "enum": [
                                                "unauthenticated"
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable; never branch on it."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional machine-readable context."
                                        }
                                    },
                                    "required": [
                                        "code",
                                        "message"
                                    ]
                                }
                            },
                            "required": [
                                "error"
                            ]
                        }
                    }
                }
            },
            "Error.not_found": {
                "description": "Not found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "enum": [
                                                "not_found"
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable; never branch on it."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Optional machine-readable context."
                                        }
                                    },
                                    "required": [
                                        "code",
                                        "message"
                                    ]
                                }
                            },
                            "required": [
                                "error"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}